src/commands/DiffCommand.js -> commands/diff/index.js
