+++ commands/diff/__tests__/__fixtures__/basic/lerna.json
