+++ commands/diff/__tests__/__fixtures__/basic/package.json
