+++ commands/diff/__tests__/__fixtures__/basic/packages/package-1/package.json
