+++ commands/diff/__tests__/diff-command.test.js
@@ -1,19 +1,19 @@
 "use strict";
 
-jest.mock("../src/ChildProcessUtilities");
-jest.mock("../src/GitUtilities");
+jest.mock("@lerna/child-process");
+jest.mock("@lerna/git-utils");
 
 const path = require("path");
 
 // mocked modules
-const ChildProcessUtilities = require("../src/ChildProcessUtilities");
-const GitUtilities = require("../src/GitUtilities");
+const ChildProcessUtilities = require("@lerna/child-process");
+const GitUtilities = require("@lerna/git-utils");
 
 // helpers
-const initFixture = require("./helpers/initFixture");
+const initFixture = require("@lerna-test/init-fixture")(__dirname);
 
 // file under test
-const lernaDiff = require("./helpers/command-runner")(require("../src/commands/DiffCommand"));
+const lernaDiff = require("@lerna-test/command-runner")(require("../command"));
 
 describe("DiffCommand", () => {
   ChildProcessUtilities.spawn.mockResolvedValue();
@@ -21,7 +21,7 @@ describe("DiffCommand", () => {
   GitUtilities.hasCommit.mockReturnValue(true);
 
   it("should diff packages from the first commit", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     GitUtilities.getFirstCommit.mockReturnValueOnce("beefcafe");
 
@@ -35,7 +35,7 @@ describe("DiffCommand", () => {
   });
 
   it("should diff packages from the most recent tag", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     GitUtilities.hasTags.mockReturnValueOnce(true);
     GitUtilities.getLastTaggedCommit.mockReturnValueOnce("cafedead");
@@ -50,7 +50,7 @@ describe("DiffCommand", () => {
   });
 
   it("should diff a specific package", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     GitUtilities.getFirstCommit.mockReturnValueOnce("deadbeef");
 
@@ -64,7 +64,7 @@ describe("DiffCommand", () => {
   });
 
   it("should error when attempting to diff a package that doesn't exist", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     try {
       await lernaDiff(testDir)("missing");
@@ -74,7 +74,7 @@ describe("DiffCommand", () => {
   });
 
   it("should error when running in a repository without commits", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     GitUtilities.hasCommit.mockReturnValueOnce(false);
 
@@ -86,7 +86,7 @@ describe("DiffCommand", () => {
   });
 
   it("should error when git diff exits non-zero", async () => {
-    const testDir = await initFixture("DiffCommand/basic");
+    const testDir = await initFixture("basic");
 
     ChildProcessUtilities.spawn.mockImplementationOnce(() => {
       const nonZero = new Error("An actual non-zero, not git diff pager SIGPIPE");
