+++ commands/diff/lib/get-last-commit.js
@@ -1,0 +1,13 @@
+"use strict";
+
+const GitUtilities = require("@lerna/git-utils");
+
+module.exports = getLastCommit;
+
+function getLastCommit(execOpts) {
+  if (GitUtilities.hasTags(execOpts)) {
+    return GitUtilities.getLastTaggedCommit(execOpts);
+  }
+
+  return GitUtilities.getFirstCommit(execOpts);
+}
