+++ commands/diff/package.json
@@ -14,7 +14,8 @@
   },
   "files": [
     "command.js",
-    "index.js"
+    "index.js",
+    "lib"
   ],
   "main": "index.js",
   "engines": {
@@ -30,5 +31,11 @@
   },
   "scripts": {
     "test": "echo \"Run tests from root\" && exit 1"
+  },
+  "dependencies": {
+    "@lerna/child-process": "file:../../core/child-process",
+    "@lerna/command": "file:../../core/command",
+    "@lerna/git-utils": "file:../../core/git-utils",
+    "@lerna/validation-error": "file:../../core/validation-error"
   }
 }
